feat: add career gap evaluation framework with golden set cases
- Introduced a new module for career gap evaluations, including a golden set of cases to define expected behaviors and outcomes.
- Implemented live report tests for evaluating career gap strategies against the golden set.
- Created live strategy functions to assess career gaps based on user profiles and project contexts.
- Added tests for generating career gap reports and comparing strategies.
- Defined types and interfaces for career gap concepts, mutations, and strategies.
- Enhanced question answering functionality to support reopening answered questions and managing project contexts.
- Updated observability types to include new gap comparison metrics and validation statuses.

+++ agent-service/app/model_policy.py
@@ -157,7 +157,7 @@ def get_gap_escalation_policy() -> GapEscalationPolicy:
         close_candidate_margin=_float(_env("AGENT_GAP_ESCALATION_MARGIN"), 0.05),
         low_confidence_threshold=_float(_env("AGENT_GAP_ESCALATION_LOW_CONFIDENCE"), 0.45),
         high_impact_threshold=_float(_env("AGENT_GAP_ESCALATION_HIGH_IMPACT"), 0.9),
-        complex_path_threshold=max(1, int(math.floor(_float(_env("AGENT_GAP_ESCALATION_COMPLEXITY"), 2, 1)))),
+        complex_path_threshold=max(1, math.floor(_float(_env("AGENT_GAP_ESCALATION_COMPLEXITY"), 2, 1))),
     )
